création du système de réservation

+++ sntp-resa/src/main/java/fr/miage/m1/sntp/dao/ReservationDao.java
@@ -7,7 +7,7 @@ import java.util.List;
 
 public interface ReservationDao {
     List<Reservation> findAll();
-    Reservation findById(int id) throws ReservationException;
+    Reservation findById(Long id) throws ReservationException;
     void save(Reservation reservation);
     void update(Reservation reservation);
     void delete(Reservation reservation);
